Only evaluate site when trait is present
